Foundation-Model-Systems-Engineering-On-Device-Multi-Modal-AI/ai-model-pipeline/requirements.txt
# Core ML & AI Frameworks
torch
transformers
accelerate
datasets

# Model Optimization & Conversion
coremltools
numpy==1.26.4

# Utilities
python-dotenv
pyyaml
tqdm
notebook